Move `RegionTracker` to `region_infer`
In terms of code organisation, this is a lot cleaner and allows
tighter access modifiers.
